discord-bot/pyproject.toml
[tool.poetry]
name = "disocrd-bot"
version = "0.1.0"
description = ""
authors = ["Maximilian Noller <maximilian.noller@uni-bielefeld.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
